% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-array.R
\name{adj_clamp}
\alias{adj_clamp}
\title{Clamp values outside the specified range}
\usage{
adj_clamp(arr, lo = -Inf, hi = Inf)
}
\arguments{
\item{arr}{array or matrix}

\item{lo}{low value.  Values lower than this will be replaced with this value.
Default: -Inf}

\item{hi}{Values higher than this will be replaced with this value. Default: Inf}
}
\value{
adjusted array
}
\description{
Clamp values outside the specified range
}
\examples{
arr <- array(1:12, c(4, 3, 1))
arr
adj_clamp(arr, 10, 20)
}
\seealso{
Other array adjustment functions: 
\code{\link{adj_gamma}()},
\code{\link{adj_infinite}()},
\code{\link{adj_rescale}()},
\code{\link{adj_shift_negatives_global}()},
\code{\link{adj_shift_negatives_local}()}
}
\concept{array adjustment functions}
