% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-array.R
\name{adj_gamma}
\alias{adj_gamma}
\title{Adjust gamma}
\usage{
adj_gamma(arr, gamma = 1/2.2)
}
\arguments{
\item{arr}{array or matrix}

\item{gamma}{gamma correction factor. Default: 1/2.2}
}
\value{
adjusted array
}
\description{
Adjust gamma
}
\examples{
arr <- array(1:12, c(4, 3, 1))
arr
adj_gamma(arr)
}
\seealso{
Other array adjustment functions: 
\code{\link{adj_clamp}()},
\code{\link{adj_infinite}()},
\code{\link{adj_rescale}()},
\code{\link{adj_shift_negatives_global}()},
\code{\link{adj_shift_negatives_local}()}
}
\concept{array adjustment functions}
