% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-array.R
\name{adj_infinite}
\alias{adj_infinite}
\title{Replace infinite values with the minimum/maximum of the finite values}
\usage{
adj_infinite(arr)
}
\arguments{
\item{arr}{array or matrix}
}
\value{
adjusted array
}
\description{
Replace infinite values with the minimum/maximum of the finite values
}
\examples{
arr <- array(c(-Inf, Inf, 1:10), c(4, 3, 1))
arr
adj_infinite(arr)
}
\seealso{
Other array adjustment functions: 
\code{\link{adj_clamp}()},
\code{\link{adj_gamma}()},
\code{\link{adj_rescale}()},
\code{\link{adj_shift_negatives_global}()},
\code{\link{adj_shift_negatives_local}()}
}
\concept{array adjustment functions}
