% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-array.R
\name{adj_rescale}
\alias{adj_rescale}
\title{Linearly rescale values to lie between the given limits}
\usage{
adj_rescale(arr, lo, hi)
}
\arguments{
\item{arr}{array or matrix}

\item{lo, hi}{limits}
}
\value{
adjusted array
}
\description{
Infinite values will be clamped to the limits
}
\examples{
arr <- array(1:24, c(4, 3, 2))
arr
adj_rescale(arr, 0, 1)
}
\seealso{
Other array adjustment functions: 
\code{\link{adj_clamp}()},
\code{\link{adj_gamma}()},
\code{\link{adj_infinite}()},
\code{\link{adj_shift_negatives_global}()},
\code{\link{adj_shift_negatives_local}()}
}
\concept{array adjustment functions}
