% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{array_to_df}
\alias{array_to_df}
\title{Convert array to a linear data.frame. Preserves array names if present.}
\usage{
array_to_df(arr)
}
\arguments{
\item{arr}{array}
}
\value{
data.frame with 'x', 'y', 'z', 'channel' and 'value.'  'channel' will 
        be the channel name if found, otherwise it is equivalent to 'z'.
}
\description{
This conversion is useful when preparing the data to summarise with
\code{ggplot}.
}
\examples{
arr <- array(1:24, dim = c(4, 3, 2))
array_to_df(arr)
}
