% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr-info.R
\name{exr_info}
\alias{exr_info}
\title{Extract the metadata from an EXR file}
\usage{
exr_info(filename, verbosity = 0)
}
\arguments{
\item{filename}{EXR filename or connection}

\item{verbosity}{verbosity. Default: 0}
}
\value{
Named list of image attributes
}
\description{
This will extract attributes from any of EXR file.
}
\examples{
filename <- system.file("image/rstats.exr", package = "picohdr")
exr_info(filename)
}
