% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr-types.R
\docType{data}
\name{exr_type}
\alias{exr_type}
\title{Functions for creating valid EXR type objects}
\format{
An object of class \code{list} of length 23.
}
\usage{
exr_type
}
\description{
This is a list of functions for creating EXR objects of a particular
EXR type.  Each function does checks for argument validity and calculates
size information required for EXR output.
}
\details{
Refer to official OpenEXR documentation
}
\examples{
# Create a v2f type
exr_type$v2f(c(12.1, 2.3))

# Create an attribute
exr_attrs(copyright = exr_type$string("mike"))
}
\keyword{datasets}
