% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.array}
\alias{plot.array}
\title{Plot method for matrices and arrays}
\usage{
\method{plot}{array}(x, interpolate = TRUE, ...)
}
\arguments{
\item{x}{matrix or array}

\item{interpolate}{Default: TRUE}

\item{...}{other arguments passed to plot()}
}
\value{
None.
}
\description{
Plot method for matrices and arrays
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
filename <- system.file("image/rstats.pfm.bz2", package = "picohdr")
image <- read_pfm(filename)
image <- adj_gamma(image)
plot(image)
\dontshow{\}) # examplesIf}
}
