% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.exr_type}
\alias{print.exr_type}
\title{Print 'exr_type' objects}
\usage{
\method{print}{exr_type}(x, ...)
}
\arguments{
\item{x}{exr_type object}

\item{...}{other arguments passed on to NextMethod}
}
\value{
None
}
\description{
Print 'exr_type' objects
}
\examples{
bbox <- exr_type$box2i(0, 0, 1, 1)
print(bbox)
}
